requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai>=0.4.0
python-multipart==0.0.6
pydantic==2.5.0
psycopg2-binary==2.9.9
Pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.16.3
authlib==1.3.0
itsdangerous==2.1.2
google-cloud-aiplatform==1.38.0
